requirements.txt
flask
flask-restx
flask-cors
flask-sqlalchemy
flask-bcrypt
pyjwt
pytest
psycopg2-binary
gunicorn
flake8
black
isort